remove duplicate implementations
