+++ webapi/src/main/scala/org/knora/webapi/messages/util/search/gravsearch/prequery/GravsearchToCountPrequeryTransformer.scala
@@ -9,7 +9,6 @@ import zio.Task
 import zio.ZIO
 
 import org.knora.webapi.ApiV2Schema
-import org.knora.webapi.messages.SmartIri
 import org.knora.webapi.messages.util.search._
 import org.knora.webapi.messages.util.search.gravsearch.types.GravsearchTypeInspectionResult
 
@@ -32,30 +31,6 @@ class GravsearchToCountPrequeryTransformer(
       querySchema = querySchema
     ) {
 
-  override def transformStatementInWhere(
-    statementPattern: StatementPattern,
-    inputOrderBy: Seq[OrderCriterion],
-    limitInferenceToOntologies: Option[Set[SmartIri]] = None
-  ): Task[Seq[QueryPattern]] =
-    // Include any statements needed to meet the user's search criteria, but not statements that would be needed for permission checking or
-    // other information about the matching resources or values.
-    processStatementPatternFromWhereClause(
-      statementPattern = statementPattern,
-      inputOrderBy = inputOrderBy
-    )
-
-  override def transformFilter(filterPattern: FilterPattern): Task[Seq[QueryPattern]] = ZIO.attempt {
-    val filterExpression: TransformedFilterPattern =
-      transformFilterPattern(filterPattern.expression, typeInspectionResult = typeInspectionResult, isTopLevel = true)
-
-    filterExpression.expression match {
-      case Some(expression: Expression) => filterExpression.additionalPatterns :+ FilterPattern(expression)
-
-      case None => filterExpression.additionalPatterns // no FILTER expression given
-    }
-
-  }
-
   override def getSelectColumns: Task[Seq[SelectQueryColumn]] =
     // return count aggregation function for main variable
     ZIO.succeed(Seq(Count(inputVariable = mainResourceVariable, distinct = true, outputVariableName = "count")))
@@ -72,17 +47,4 @@ class GravsearchToCountPrequeryTransformer(
   override def getOffset(inputQueryOffset: Long, limit: Int): Task[Long] =
     // count queries do not consider offsets since there is only one result row
     ZIO.succeed(0L)
-
-  override def optimiseQueryPatterns(patterns: Seq[QueryPattern]): Task[Seq[QueryPattern]] =
-    ZIO.attempt(
-      GravsearchQueryOptimisationFactory
-        .getGravsearchQueryOptimisationFeature(
-          typeInspectionResult = typeInspectionResult,
-          querySchema = querySchema
-        )
-        .optimiseQueryPatterns(patterns)
-    )
-
-  override def transformLuceneQueryPattern(luceneQueryPattern: LuceneQueryPattern): Task[Seq[QueryPattern]] =
-    ZIO.succeed(Seq(luceneQueryPattern))
 }
